+++ crates/common/src/network/mta.rs
@@ -203,7 +203,7 @@ impl Server {
     }
 
     pub fn get_trusted_sieve_script(&self, name: &str, session_id: u64) -> Option<&Arc<Sieve>> {
-        self.core.sieve.trusted_scripts.get(name).or_else(|| {
+        self.core.sieve.trusted_script(name).or_else(|| {
             trc::event!(
                 Sieve(trc::SieveEvent::ScriptNotFound),
                 Id = name.to_string(),
@@ -215,7 +215,7 @@ impl Server {
     }
 
     pub fn get_untrusted_sieve_script(&self, name: &str, session_id: u64) -> Option<&Arc<Sieve>> {
-        self.core.sieve.untrusted_scripts.get(name).or_else(|| {
+        self.core.sieve.untrusted_script(name).or_else(|| {
             trc::event!(
                 Sieve(trc::SieveEvent::ScriptNotFound),
                 Id = name.to_string(),